:sparkles: add announcement to contest
