+++ src/context/AuthContext.tsx
@@ -5,9 +5,10 @@ import { cache } from 'swr'
 
 import { useHttp } from './HttpContext'
 
-import { useDisclosure, useForceUpdate } from '@chakra-ui/hooks'
+import { useDisclosure } from '@chakra-ui/hooks'
 
 import { LoginModal } from '@src/components/Login'
+import { useForceUpdate } from '@src/hooks/useForceUpdate'
 import { AuthRes, LoginReq, User } from '@src/user/types'
 
 export interface AuthProviderProps {
